// src/apis/model/farmerDetails.ts
/**
 * Generated by orval v6.14.3 🍺
 * Do not edit manually.
 * Farmer App API
 * OpenAPI spec version: v1
 */
import type { FarmerPlan } from './farmerPlan';

export interface FarmerDetails {
  displayName?: string | null;
  emailId?: string | null;
  mobileNumber?: string | null;
  farmerId?: string | null;
  fcmToken?: string | null;
  signInProvider?: string | null;
  allowedArea?: number;
  farmerPlans?: FarmerPlan;
  farmonautAllowedCrops?: number;
  farmonautCropsConsumed?: number;
  farmonautCropsAvailable?: number;
}